devtools/requirements.txt
click==8.1.7
pyjwt==2.8.0
cryptography==42.0.5